refactored params cast to ints

+++ src/Main/GameObject.java
@@ -4,11 +4,11 @@ import java.awt.*;
 
 public abstract class GameObject {
 
-    protected int x, y;
+    protected float x, y;
     protected ID id;
     protected int velX, velY;
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
@@ -26,11 +26,11 @@ public abstract class GameObject {
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
